+++ tortoise/backends/asyncpg/executor.py
@@ -15,15 +15,16 @@ class AsyncpgExecutor(BaseExecutor):
     def Parameter(self, pos: int) -> Parameter:
         return Parameter("$%d" % (pos + 1,))
 
-    def _prepare_insert_statement(self, columns: List[str]) -> str:
+    def _prepare_insert_statement(self, columns: List[str], no_generated: bool = False) -> str:
         query = (
             self.db.query_class.into(self.model._meta.basetable)
             .columns(*columns)
             .insert(*[self.Parameter(i) for i in range(len(columns))])
         )
-        generated_fields = self.model._meta.generated_db_fields
-        if generated_fields:
-            query = query.returning(*generated_fields)
+        if not no_generated:
+            generated_fields = self.model._meta.generated_db_fields
+            if generated_fields:
+                query = query.returning(*generated_fields)
         return str(query)
 
     async def _process_insert_result(self, instance: Model, results: Optional[asyncpg.Record]):